requirements.txt
alembic==1.11.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
qrcode==7.4.2
Pillow==10.1.0
reportlab==4.0.7
openpyxl==3.1.5
python-dotenv==1.0.1

source venv/bin/activate
python app.py
twilio==8.4.1
twilio>=8.0.0
